+++ awesome-m3-themes/connected/.config/nvim/lua/colors.lua
@@ -1,0 +1,31 @@
+-- theme connected
+-- base colors: https://github.com/catppuccin/nvim/blob/main/lua/catppuccin/palettes/mocha.lua
+-- WGAG AAA need a contrast of 7:1 for normal text
+return {
+	rosewater = "#f4f1f1", -- white light
+	flamingo = "#f4f2f1", -- white
+	pink = "#fac7d5", -- magenta light
+	mauve = "#ec7998", -- magenta
+	red = "#ec897a", -- red
+	maroon = "#f49d90", -- red light
+	peach = "#f2e7a7", -- yellow
+	yellow = "#f7f0c6", -- yellow light
+	green = "#87eea5", -- green
+	teal = "#9df6bc", -- green light
+	sky = "#d6fbff", -- cyan light
+	sapphire = "#87eeff", -- cyan
+	blue = "#6ceabf", -- blue
+	lavender = "#b9f8e3", -- blue light
+	text = "#E2E3DF",
+	--subtext1 = '#',
+	--subtext0 = '#',
+	overlay2 = "#7b9d8a",
+	-- ...
+	--surface2 = '#373A37',
+	--surface1 = '#333533',
+	surface0 = "#282B28",
+	base = "#111412",
+	mantle = "#1A1C1A",
+	--crust = '#0C0F0D',
+	--crust = '#0E0E13', -- deep dark
+}