+++ awesome-m3-themes/focus/.config/nvim/lua/colors.lua
@@ -1,0 +1,31 @@
+-- theme focus
+-- base colors: https://github.com/catppuccin/nvim/blob/main/lua/catppuccin/palettes/mocha.lua
+-- WGAG AAA need a contrast of 7:1 for normal text
+return {
+    rosewater = '#d1d6dc', -- white light
+    flamingo = '#9ea7b3', -- white
+    pink = '#f1d0ed', -- magenta light
+    mauve = '#d88dce', -- magenta
+    red = '#fa2e67', -- red
+    maroon = '#FF85a7', -- red light
+    peach = '#e650a0', -- yellow
+    yellow = '#ffa3be', -- yellow light
+    green = '#9fedea', -- green
+    teal = '#9fedff', -- green light
+    sky = '#54E0C5', -- cyan light
+    sapphire = '#25D6D0', -- cyan
+    blue = '#8596d1', -- blue
+    lavender = '#c6ceec', -- blue light
+    text = '#E2E2E2',
+    --subtext1 = '#',
+    --subtext0 = '#',
+    overlay2 = '#755892',
+    -- ...
+    --surface2 = '#393939',
+    --surface1 = '#353535',
+    surface0 = '#2A2A2A',
+    base = '#131313',
+    mantle = '#1B1B1B',
+    --crust = '#0E0E0E',
+    --crust = '#0E0E13', -- deep dark
+}