+++ awesome-m3-themes/lines/.config/nvim/lua/colors.lua
@@ -1,0 +1,31 @@
+-- theme lines
+-- base colors: https://github.com/catppuccin/nvim/blob/main/lua/catppuccin/palettes/mocha.lua
+-- WGAG AAA need a contrast of 7:1 for normal text
+return {
+    rosewater = '#f4f1f3', -- white light
+    flamingo = '#f4f1f3', -- white
+    pink = "#f2cff2", -- magenta light
+    mauve = "#ea8bda", -- magenta
+    red = "#e72ea7", -- red
+    maroon = "#f94db9", -- red light
+    peach = "#f37e4e", -- yellow
+    yellow = "#ffad4f", -- yellow light
+    green = "#6c92cb", -- green
+    teal = "#adc4e6", -- green light
+    sky = "#dd89fd", -- cyan light
+    sapphire = "#c471e2", -- cyan
+    blue = "#a47bfb", -- blue
+    lavender = "#cacce7", -- blue light
+    text = '#E2E2E9',
+    --subtext1 = '#',
+    --subtext0 = '#',
+    overlay2 = '#9e9e9e',
+    -- ...
+    --surface2 = '#38393F',
+    --surface1 = '#33343A',
+    surface0 = '#282A2F',
+    base = '#121318',
+    mantle = '#1A1B21',
+    --crust = '#0D0E13',
+    --crust = '#0E0E13', -- deep dark
+}