+++ awesome-m3-themes/miami/.config/nvim/lua/colors.lua
@@ -1,0 +1,31 @@
+-- theme miami
+-- base colors: https://github.com/catppuccin/nvim/blob/main/lua/catppuccin/palettes/mocha.lua
+-- WGAG AAA need a contrast of 7:1 for normal text
+return {
+    rosewater = '#efefef', -- white light
+    flamingo = '#d1b1af', -- white
+    pink = "#bda7bf", -- magenta light
+    mauve = "#cd98cd", -- magenta
+    red = "#e55958", -- red
+    maroon = "#f17675", -- red light
+    peach = "#d2A452", -- yellow
+    yellow = "#e3c28c", -- yellow light
+    green = "#d3b793", -- green
+    teal = "#d3b693", -- green light
+    sky = "#dfe9f1", -- cyan light
+    sapphire = "#9fbfc7", -- cyan
+    blue = "#a7a4c2", -- blue
+    lavender = "#a6a3c3", -- blue light
+    text = '#F0DEDD',
+    --subtext1 = '#',
+    --subtext0 = '#',
+    overlay2 = '#895d84',
+    -- ...
+    --surface2 = '#423736',
+    --surface1 = '#3D3232',
+    surface0 = '#322827',
+    base = '#1A1111',
+    mantle = '#231919',
+    --crust = '#140C0C',
+    --crust = '#0E0E13', -- deep dark
+}