+++ awesome-m3-themes/morpho/.config/nvim/lua/colors.lua
@@ -1,0 +1,31 @@
+-- theme morpho
+-- base colors: https://github.com/catppuccin/nvim/blob/main/lua/catppuccin/palettes/mocha.lua
+-- WGAG AAA need a contrast of 7:1 for normal text
+return {
+    rosewater = '#f1f1f4', -- white light
+    flamingo = '#bbc3e4', -- white
+    pink = "#fdfaff", -- magenta light
+    mauve = "#daaaf3", -- magenta
+    red = "#d075eb", -- red
+    maroon = "#edc3f9", -- red light
+    peach = "#90e0ef", -- yellow
+    yellow = "#dff7fc", -- yellow light
+    green = "#9cf1c6", -- green
+    teal = "#d0fbea", -- green light
+    sky = "#aad8fa", -- cyan light
+    sapphire = "#69b9ea", -- cyan
+    blue = "#7879f1", -- blue
+    lavender = "#9192f7", -- blue light
+    text = '#E6E0EA',
+    --subtext1 = '#',
+    --subtext0 = '#',
+    overlay2 = '#6f60a6',
+    -- ...
+    --surface2 = '#3A3840',
+    --surface1 = '#36343B',
+    surface0 = '#2B2930',
+    base = '#141219',
+    mantle = '#1C1B21',
+    --crust = '#0F0D14',
+    --crust = '#0E0E13', -- deep dark
+}