+++ awesome-m3-themes/sci/.config/nvim/lua/colors.lua
@@ -1,0 +1,31 @@
+-- theme sci
+-- base colors: https://github.com/catppuccin/nvim/blob/main/lua/catppuccin/palettes/mocha.lua
+-- WGAG AAA need a contrast of 7:1 for normal text
+return {
+    rosewater = '#f1f3f4', -- white light
+    flamingo = '#d3ddde', -- white
+    pink = "#d1e1f0", -- magenta light
+    mauve = "#8fb4d6", -- magenta
+    red = "#fb6c67", -- red
+    maroon = "#fe8378", -- red light
+    peach = "#d39e4e", -- yellow
+    yellow = "#dfbd4f", -- yellow light
+    green = "#6acdba", -- green
+    teal = "#7eddca", -- green light
+    sky = "#7ea0dd", -- cyan light
+    sapphire = "#6a8ecd", -- cyan
+    blue = "#68b0cf", -- blue
+    lavender = "#7cc0de", -- blue light
+    text = '#DFE3E5',
+    --subtext1 = '#',
+    --subtext0 = '#',
+    overlay2 = '#627f84',
+    -- ...
+    --surface2 = '#353A3B',
+    --surface1 = '#313637',
+    surface0 = '#262B2C',
+    base = '#0F1416',
+    mantle = '#181C1E',
+    --crust = '#0A0F10',
+    --crust = '#0E0E13', -- deep dark
+}