+++ neovim/.config/nvim/lua/ui/colorizer.lua
@@ -1,0 +1,11 @@
+return {
+    'brenoprata10/nvim-highlight-colors',
+    event = { 'BufReadPost' },
+    cmd = 'HighlightColors',
+    opts = {
+        render = 'background',
+        enable_tailwind = true,
+        enable_hex = true,
+        virtual_symbol = '󱓻',
+    },
+}