+++ neovim/.config/nvim/lua/utils/icon.lua
@@ -1,0 +1,14 @@
+local M = {}
+
+function M.get(name)
+    local i = {
+        Diagnostic = '󰒡',
+        DiagnosticError = '',
+        DiagnosticHint = '󰌵',
+        DiagnosticInfo = '󰋼',
+        DiagnosticWarn = '',
+    }
+    return i[name]
+end
+
+return M